Also consider key 'timestamp' for the time information
By default, the log event processor only uses '@timestamp' key to get
the time of the event. This commit introduces changes to also consider
key 'timestamp' as a fallback if '@timestamp' key is not present.

This is the case for the Python's 'kafka-logging-handler' library which
sets the 'timestamp' key instead of '@timestamp'