[NCL-8159] Add support to send output to different partitions
This is done while preserving message ordering for a specific build
(which has a specific process context mdc value) by sending those
messages to the same partition. It is performed by setting the key of
the output message (if not already set) to the process context mdc
value.

If the key is not set, then we set the key to a Default value. This
ensures that any messages with no keys are sent to the same partition.

This allows apps like Bifrost to run multiple instances to consume
messages from log-event-duration by having multiple partitions to
consume the data.