Fix old timestamps being sent to output queue
We sometimes experience a catastrophic error when we try to consume an
old message with an old timestamp and upon sending it to the downstream
queue, it fails with:

```
error sending message due to timestamp
timestamp with offset out of range
```

I believe this is caused in our `MergeTransformer.java` when using:
```
context.forward(...)
```

From the documentation: https://kafka.apache.org/0110/documentation/streams/core-concepts#streams_time

> Finally, whenever a Kafka Streams application writes records to Kafka, then it will also assign timestamps to these new records. The way the timestamps are assigned depends on the context:
>
>    When new output records are generated via processing some input record, for example, context.forward() triggered in the process() function call, output record timestamps are inherited from input record timestamps directly.

This commit tries to set the latest timestamp when we use
`context.forward` when processing older messages to workaround this
issue. The timestamp in the JSON payload stays the same.